New environment, updates, and documentation
- Renamed WaypointTest to WaypointForGate
- Created new environment Trap, renamed to RatEnvironment
- Added new environments to Room
- Refined mentioned environments and added better documentation and comments
